services/api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = "API service for XO game"
authors = ["chiselko6 <alex.limontov@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.63.0"
sqlalchemy = "^1.4.4"
pydantic = "^1.8.1"
alembic = "^1.5.8"
psycopg2 = "^2.8.6"
python-jose = "^3.2.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.5"
uvicorn = {extras = ["standard"], version = "^0.13.4"}
aiohttp = "^3.7.4"


[tool.poetry.dev-dependencies]
black = "^20.8b1"
isort = "^5.8.0"
mypy = "^0.812"
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
